% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\encoding{UTF-8}
\name{checking}
\alias{checking}
\title{Check for validity of parameters of the \code{preprocesss}-Funktion}
\usage{
checking(
  dta,
  var_vec,
  density_var,
  sample_weights,
  bin_width,
  bin_number,
  values_discrete,
  weights_discrete,
  domain_continuous
)
}
\arguments{
\item{dta}{Has to be of type \code{data.frame} or \code{data.table} with observations of the density variable and covariates as well as optional sampling weights for each observation.}

\item{var_vec}{Vector of variables of \code{dta} on which the covariate combinations are based. The vector has either to contain the variable names as strings or the column positions of the respective variables in \code{dta}.}

\item{density_var}{Variable of \code{dta} which corresponds to the domain of the density. Variable name as string or numeric column position of the variable in \code{dta}.}

\item{sample_weights}{Variable of \code{dta} which contains a sample weight for each observation. Variable name as string or numeric column position of the variable in \code{dta}. If missing (\code{NULL}), no sample weights are included per default.}

\item{bin_width}{Width of histogram bins partitioning the continuous domain \eqn{I}. Scalar for equidistant bins or vector with different increments. The union of the resulting bins must exactly match the domain of the continuous component. Alternative to \code{bin_number}.}

\item{bin_number}{Number of equidistant histogram bins partitioning the continuous domain \eqn{I}. Alternative to \code{bin_width}. If \code{bin_number} and \code{bin_width} are both given and the two values are not compatible, an error is returned. If neither parameter is specified, \code{bin_number=100} is used as default.}

\item{values_discrete}{Vector of values in the domain of the density that have positive probability mass (dirac measure). Defaults to missing (\code{NULL}) in which case it is set to \code{c(0, 1)}. Can also be set to be \code{FALSE} in which case the discrete component is considered to be empty, i.e., the Lebesgue measure is used as reference measure. Cannot be \code{FALSE} if \code{domain_continuous=FALSE}.}

\item{weights_discrete}{Vector of weights which have to be positive for the dirac measures corresponding to values_discrete. If missing (\code{NULL}) it is set to 1 in all components as default. Can be a scalar for equal weights for all discrete values or a vector with specific weights for each corresponding discrete value.}

\item{domain_continuous}{An interval (i.e., a vector of length 2) specifying the domain of the continuous component of the density. If missing (\code{NULL}) it is set to \code{c(0, 1)} as default. Can also be set to be \code{FALSE} in which case the discrete component is considered to be empty, i.e., a sum of dirac measures is used as reference measure. Cannot be \code{FALSE} if \code{values_discrete=FALSE}.}
}
\description{
\code{checking} checks for valid parameters for \code{preprocess}. Returns an error if any parameters requirement is violated.
}
