% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondDensReg-package.R
\docType{package}
\name{CondDensReg-package}
\alias{CondDensReg}
\alias{CondDensReg-package}
\alias{CondDensReg_package}
\alias{package-CondDensReg}
\title{CondDensReg: Conditional density regression for individual-level data}
\description{
Structured additive regression models for mixed densities as response and
scalar covariates are fitted via Poisson regression models, where the
continuous part of the density is approximated with histograms.
}
\details{
This package is intended to fit regression models with densities in mixed
(continuous/discrete) Bayes Hilbert spaces (including continuous and discrete
ones as special cases) as response given scalar covariates, based on observed
samples of the conditional distributions via a (penalized) maximum likelihood
approach. The (penalized) log-likelihood function is approximated via the
(penalized) log-likelihood of an appropriate poisson regression model, which
is then fitted using \code{mgcv}'s \code{\link[mgcv]{gam}} with a new smooth for mixed
densities. For details see Maier et al. (2023).

The main fitting function is \code{\link{dens_reg}}. Besides the (individual-
level) data, it takes a variety of arguments related to specification of
the underlying mixed Bayes Hilbert space, histogram construction for its
continuous part, and partial effects (including penalties). We also provide
\code{plot} and \code{predict} methods for the resulting fitted object.
}
\references{
Maier, E. M., Fottner, A., Stoecker, A., Okhrin, Y., & Greven, S. (2023):
Conditional density regression for individual-level data.
}
\seealso{
\code{\link{dens_reg}} for the main fitting function.
}
\author{
Eva-Maria Maier, Lea Runge, and Alexander Fottner
}
