% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms.R
\encoding{UTF-8}
\name{interactive_plots}
\alias{interactive_plots}
\title{Interactive application for plots for all covariate combinations}
\usage{
interactive_plots(
  dta,
  selected_groups = unique(dta$group_id),
  case = "all",
  abs = FALSE,
  main = NULL,
  automatic_main = TRUE,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  col_discrete = "red",
  hist = TRUE,
  ...
)
}
\arguments{
\item{dta}{\code{data.table}-object which is also an object of the sub-class \code{histogram_count_data}, i.e., the output of the \code{\link[=preprocess]{preprocess()}}-function with multiple covariate combinations, i.e. \code{group_id}s, contained.}

\item{selected_groups}{Numeric vector of \code{group_id}s for which the respective plots will be created and included in the interactive plot. If missing (\code{NULL}) plots of all groups within the range of the minimum and maximum \code{group_id} will be included.}

\item{case}{Plotting case as a \code{character}-object: can be specified as "all", "continuous" or "discrete". Determines the components of the data which will be plotted: "all" leads to a plot of the complete data, "continuous" only plots the bins displaying data of the continuous domain \eqn{I} and "discrete" shows only bins displaying the counts of the respective discrete values in \eqn{D}. Please note that the plotting case is not the same as the specification of the case in \code{\link[=preprocess]{preprocess()}} where the bins and discrete values are defined. If the case is not "all" and \code{hist} is TRUE, the sum of the areas under the bins and the discrete bars is not one. If missing (\code{NULL}), the default is "all".}

\item{abs}{\code{logical} which determines if the resulting plot shows the absolute, (weighted) counts (\code{abs=TRUE}) or the relative frequencies of the weighed counts (\code{abs=FALSE}). If missing (\code{NULL}), the default is \code{FALSE}.}

\item{main}{Optional \code{character} passing a user-defined main title for the resulting plot. If missing (\code{NULL}), the main will be generated automatically, see also \code{automatic_main}.}

\item{automatic_main}{\code{logical} determining the automatic generation of a main title for the plot which contains the number of bins, the plot case defined by \code{case}, the values of the respective covariate combination and the group_id. If \code{TRUE} and a user-defined \code{main} is passed, the automatic main is overwriting the passed main. If missing (\code{NULL}), the default is \code{TRUE}.}

\item{ylim}{An interval (i.e., a \code{numeric} vector of length 2) specifying the limits of the plot's y-axis. If missing (\code{NULL}), the limits are set to 0 and the maximal height of a bar or histogram bin.}

\item{xlab}{Optional \code{character} passing a user-defined title of the x-axis. If missing (\code{NULL}), the axis title is set to be the name of the density variable of \code{dta}.}

\item{ylab}{Optional \code{character} passing a user-defined title of the y-axis. If missing (\code{NULL}), the axis title contains the information if absolute values or frequencies are shown in a histogram (for the continuous components) or a bar plot.}

\item{col_discrete}{Name or hexadecimal RGB triplet of a color contained in base R. Determines the color of the bars of the discrete values. If missing (\code{NULL}), the default color is "red". It is recommended to use a color contrasting the black outlines of the bins of the contunuous domain especially if one discrete value lies exactly at a bin border.}

\item{hist}{\code{logical} which determines if the resulting plot is a histogram or a simple bar plot of the data. If \code{case="all"} and \code{hist=TRUE}, the areas under the bins and the height of discrete bars times the respective weight of the discrete dirac measure sum up to one. Otherwise the height of the bins is directly referring to the frequency or the (weighted) counts of the respective bin without taking the bin width or the dirac measure into account. If missing (\code{NULL}), the default is \code{TRUE}.}

\item{...}{Other parameters which can be passed to the respective plotting function specifying plotting parameters.}
}
\value{
The returned object is an interactive plot object specified as explained above as a histogram or bar plot of frequencies or absolute counts. The display of mixed case count (histogram) data is as considered in the work of Maier et al. (2023). If the slider contains group IDs which are not available in the data, an empty plot and a warning occure.
}
\description{
\code{interactive_plots} creates for all groups, i.e., all different covariate combinations considered in the data, a plot based on  \code{\link[=plot_single_histo]{plot_single_histo()}} and integrates them in an interactive application where the user can select a \code{group_id} with a slider. All specified parameters will be passed to \code{\link[=plot_single_histo]{plot_single_histo()}}. \code{interactive_plots} is also used as default plotting method for data of the sub-class \code{histogram_count_data}, i.e., the output of the \code{\link[=preprocess]{preprocess()}}-function.
}
\examples{
set.seed(101)

# create data where 0 and 1 are the discrete observations, values
# equal 2 are replaced below by drawing from a beta distribution

dta <-data.frame(
obs_density = sample(0:2, 100, replace = TRUE, prob = c(0.15, 0.1, 0.75)),
covariate1 = sample(c("a", "b"), 100, replace = TRUE),
covariate2 = sample(c("c", "d"), 100, replace = TRUE),
sample_weights = runif(100, 0, 2)
)
dta[which(dta$obs_density == 2), ]$obs_density <- rbeta(length(which(dta$obs_density == 2)), shape1 = 3, shape2 = 3)

# Create histogram count dataset for dta using 10 equidistant
# bins and default values for continuous domain, discrete
# values and discrete weights while considering a mixed case
# of continuous and discrete domains.

histo_data <- preprocess(
dta,
var_vec = c("covariate1", "covariate2"),
density_var = "obs_density",
sample_weights = "sample_weights",
bin_number = 10
)

# Plot a single histogram of the weighted frequencies.
# Show all observations.

interactive_plots(histo_data)

# Restrict the plot to only discrete observations
# of the groups 1-3, use as bar color blue and 
# show the absolute weighted counts instead of 
# the frequencies:

interactive_plots(
histo_data,
selected_groups= c(1,2,3),
case = "discrete",
abs = TRUE,
col_discrete = "blue"
)

# Restrict now the plot to only continuous observations
# and plot a bar plot without the histogram property and 
# with user-defined main, axis titles and axis limits:

interactive_plots(
  histo_data,
  case = "continuous",
  hist = FALSE,  
  main ="Title",  
  automatic_main = FALSE,  
  xlab = "x-axis",  
  ylab ="y-axis",  
  ylim = c(0,1)
  )

}
\references{
Maier, E. M., Fottner, A., Stöcker, A., Okhrin, Y., & Greven, S. (2023). Conditional density regression for individual-level data.
}
