% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms.R
\encoding{UTF-8}
\name{plot.histogram_count_data}
\alias{plot.histogram_count_data}
\title{Plot function for the observed (histogram) count data}
\usage{
\method{plot}{histogram_count_data}(dta)
}
\arguments{
\item{dta}{\code{data.table}-object which is an object of the sub-class \code{histogram_count_data}, i.e., the output of the \code{\link[=preprocess]{preprocess()}}-function with multiple covariate combinations, i.e. \code{group_id}s, contained.}
}
\value{
Interactive plot displaying all histograms for the contained groups where the respective \code{group_id} is selectable with a slider. The display of mixed case count (histogram) data is as considered in the work of Maier et al. (2023).
}
\description{
\code{plot.histogram_count_data} is the default plot method for data of the class \code{histogram_count_data}. It is using the function \code{interactive_plots} with its default parameters, i.e. it is creating an interactive plot of the frequencies of the respective (weighted) counts with histograms on \eqn{I\setminus D} and of the counts on \eqn{D} with simple bars where \eqn{I} is the interval of the continuous domain and \eqn{D} the set of discrete values.. In the plot the areas under the bins and the height of discrete bars times the respective weight of the discrete dirac measure sum up to one.
}
\examples{
set.seed(101)

# create data where 0 and 1 are the discrete observations, values
# equal 2 are replaced below by drawing from a beta distribution

dta <-data.frame(
obs_density = sample(0:2, 100, replace = TRUE, prob = c(0.15, 0.1, 0.75)),
covariate1 = sample(c("a", "b"), 100, replace = TRUE),
covariate2 = sample(c("c", "d"), 100, replace = TRUE),
sample_weights = runif(100, 0, 2)
)
dta[which(dta$obs_density == 2), ]$obs_density <- rbeta(length(which(dta$obs_density == 2)), shape1 = 3, shape2 = 3)

# Create histogram count dataset for dta using 10 equidistant
# bins and default values for continuous domain, discrete
# values and discrete weights while considering a mixed case
# of continuous and discrete domains.

histo_data <- preprocess(
dta,
var_vec = c("covariate1", "covariate2"),
density_var = "obs_density",
sample_weights = "sample_weights",
bin_number = 10
)

# Plot histo_data:
plot(histo_data)
}
\references{
Maier, E. M., Fottner, A., Stöcker, A., Okhrin, Y., & Greven, S. (2023). Conditional density regression for individual-level data.
}
