% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_density_smooth.R
\name{Predict.matrix.mdspline.smooth}
\alias{Predict.matrix.mdspline.smooth}
\title{Predict matrix method function for mixed density smooth}
\usage{
\method{Predict.matrix}{mdspline.smooth}(object, data)
}
\arguments{
\item{object}{a smooth specification object, usually generated by a term
\code{ti(x,bs="md",...)}}

\item{data}{ A data frame containing the values of the (named) covariates at which the smooth term is to be 
evaluated. Exact requirements are as for \code{\link[mgcv]{smooth.construct}} and \code{smooth.construct2}}
}
\value{
A matrix mapping the coeffients for the smooth term to its values at
the supplied data values.
}
\description{
\code{\link[mgcv]{Predict.matrix}} method function for smooth class
\code{mdspline.smooth} to enable prediction from a model fitted with
\code{mgcv}'s \code{\link[mgcv]{gam}}.
}
\details{
The Predict matrix function is not normally called directly, but is rather
used internally by \code{mgcv}'s \code{\link[mgcv]{predict.gam}} etc. to
predict from a fitted \code{\link[mgcv]{gam}} model. See \code{\link[mgcv]{Predict.matrix}}
for more details, or \code{\link{smooth.construct.md.smooth.spec}} for details
on the mixed density smooth \code{"md"}.
}
\references{
Maier, E. M., Fottner, A., Stoecker, A., Okhrin, Y., & Greven, S. (2023):
Conditional density regression for individual-level data.
}
\author{
Eva-Maria Maier, Alexander Fottner
}
