% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms.R
\encoding{UTF-8}
\name{save_plots}
\alias{save_plots}
\title{Saving plots of observed (histogram) count data}
\usage{
save_plots(
  dta,
  selected_groups = unique(dta$group_id),
  single = TRUE,
  path = "./Images/",
  case = "all",
  main = NULL,
  automatic_main = TRUE,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  col_discrete = "red",
  file_names = NULL,
  file_names_stem = NULL,
  plot_width = ifelse(single, 7, 480),
  plot_height = ifelse(single, 7, 480),
  abs = FALSE,
  hist = TRUE,
  ...
)
}
\arguments{
\item{dta}{\code{data.table}-object which is also an object of the sub-class \code{histogram_count_data}, i.e., the output of the \code{\link[=preprocess]{preprocess()}}-function. Subset in such a way, that only one covariate combination, i.e. one \code{group_id} is contained.}

\item{selected_groups}{Numeric vector of \code{group_id}s for which the respective plots will be created and saved. If missing (\code{NULL}) plots of all groups within the range of the minimum and maximum \code{group_id} will be saved.}

\item{single}{\code{logical} indicating if all plots are saved as single PDFs if \code{single=TRUE} or jointly in one document with four plots per page if \code{single=FALSE}. If missing (\code{NULL}), the default is \code{TRUE}.}

\item{path}{Directory path where resulting plots will be saved. If not specified by the user then the storage directory is the subdirectory "Images" of the current working directory. If this does not yet exist, it will be created.}

\item{case}{Plotting case as a \code{character}-object: can be specified as "all", "continuous" or "discrete". Determines the components of the data which will be plotted: "all" leads to a plot of the complete data, "continuous" only plots the bins displaying data of the continuous domain \eqn{I} and "discrete" shows only bins displaying the counts of the respective discrete values in \eqn{D}. Please note that the plotting case is not the same as the specification of the case in \code{\link[=preprocess]{preprocess()}} where the bins and discrete values are defined. If the case is not "all" and \code{hist} is TRUE, the sum of the areas under the bins and the discrete bars is not one. If missing (\code{NULL}), the default is "all".}

\item{main}{Optional \code{character} passing a user-defined main title for the resulting plot. If missing (\code{NULL}), the main will be generated automatically, see also \code{automatic_main}.}

\item{automatic_main}{\code{logical} determining the automatic generation of a main title for the plot which contains the number of bins, the plot case defined by \code{case}, the values of the respective covariate combination and the group_id. If \code{TRUE} and a user-defined \code{main} is passed, the automatic main is overwriting the passed main. If missing (\code{NULL}), the default is \code{TRUE}.}

\item{ylim}{An interval (i.e., a \code{numeric} vector of length 2) specifying the limits of the plot's y-axis. If missing (\code{NULL}), the limits are set to 0 and the maximal height of a bar or histogram bin.}

\item{xlab}{Optional \code{character} passing a user-defined title of the x-axis. If missing (\code{NULL}), the axis title is set to be the name of the density variable of \code{dta}.}

\item{ylab}{Optional \code{character} passing a user-defined title of the y-axis. If missing (\code{NULL}), the axis title contains the information if absolute values or frequencies are shown in a histogram (for the continuous components) or a bar plot.}

\item{col_discrete}{Name or hexadecimal RGB triplet of a color contained in base R. Determines the color of the bars of the discrete values. If missing (\code{NULL}), the default color is "red". It is recommended to use a color contrasting the black outlines of the bins of the contunuous domain especially if one discrete value lies exactly at a bin border.}

\item{file_names}{Vector of file names can be passed by the user. If \code{single=TRUE} the number of file names has to be the same as the number of selected groups. If \code{single=FALSE} only one file name can be passed. If missing (\code{NULL}), the names of the single plots will be their \code{group_id} with the prefix "hist_". The name of the file if \code{single=FALSE} is "all.pdf". A warning occurs if \code{file_names_stem} is not \code{NULL} because only one of the two parameters \code{file_names_stem} and \code{file_names} can be specified.}

\item{file_names_stem}{\code{character} which replaces the prefix "hist_" in the automatically generated file name if \code{file_names=NULL}. If missing (\code{NULL}), the names of the single plots will be their \code{group_id} with the prefix "hist_". The name of the file if \code{single=FALSE} is "all.pdf". A warning occurs if \code{file_names} is not \code{NULL} because only one of the two parameters \code{file_names_stem} and \code{file_names} can be specified.}

\item{plot_width}{Value for the plot width in the case of \code{single=TRUE}. For \code{single=FALSE} the plot width is not adjustable. If missing (\code{NULL}), the default value if \code{single=TRUE} is as for the \code{pdf()}-function 7 inches.}

\item{plot_height}{Value for the plot height in the case of \code{single=TRUE}. For \code{single=FALSE} the plot height is not adjustable. If missing (\code{NULL}), the default value if \code{single=TRUE} is as for the \code{pdf()}-function 7 inches.}

\item{abs}{\code{logical} which determines if the resulting plot shows the absolute, (weighted) counts (\code{abs=TRUE}) or the relative frequencies of the weighed counts (\code{abs=FALSE}). If missing (\code{NULL}), the default is \code{FALSE}.}

\item{hist}{\code{logical} which determines if the resulting plot is a histogram or a simple bar plot of the data. If \code{case="all"} and \code{hist=TRUE}, the areas under the bins and the height of discrete bars times the respective weight of the discrete dirac measure sum up to one. Otherwise the height of the bins is directly referring to the frequency or the (weighted) counts of the respective bin without taking the bin width or the dirac measure into account. If missing (\code{NULL}), the default is \code{TRUE}.}

\item{...}{Other parameters which can be passed to the respective plotting function specifying plotting parameters}
}
\value{
Saved plots in PDF-format.
}
\description{
\code{save_plots} offers the possibility to create plots for selected groups with \code{\link[=plot_single_histo]{plot_single_histo()}} and then save them locally as pdf.
}
\examples{
set.seed(101)

# create data where 0 and 1 are the discrete observations, values
# equal 2 are replaced below by drawing from a beta distribution

dta <-data.frame(
obs_density = sample(0:2, 100, replace = TRUE, prob = c(0.15, 0.1, 0.75)),
covariate1 = sample(c("a", "b"), 100, replace = TRUE),
covariate2 = sample(c("c", "d"), 100, replace = TRUE),
sample_weights = runif(100, 0, 2)
)
dta[which(dta$obs_density == 2), ]$obs_density <- rbeta(length(which(dta$obs_density == 2)), shape1 = 3, shape2 = 3)

# Create histogram count dataset for dta using 10 equidistant
# bins and default values for continuous domain, discrete
# values and discrete weights while considering a mixed case
# of continuous and discrete domains.

histo_data <- preprocess(
dta,
var_vec = c("covariate1", "covariate2"),
density_var = "obs_density",
sample_weights = "sample_weights",
bin_number = 10
)

# Plot a single histogram of the weighted frequencies.
# Show all observations and save all plots in single PDFs.

save_plots(histo_data)

# Restrict the plot to only discrete observations
# of the groups 1-3, use as bar color blue and 
# show the absolute weighted counts instead of 
# the frequencies. Name the three resulting PDF
# with user-defined names.

save_plots(
histo_data,
selected_groups= c(1,2,3),
case = "discrete",
abs = TRUE,
col_discrete = "blue",
file_names= c("one","two","three")
)

# Restrict now the plot to only continuous observations
# and plot a bar plot without the histogram property and 
# with user-defined main, axis titles and axis limits.
# Save all plots in one PDF.

save_plots(
  histo_data,
  case = "continuous",
  hist = FALSE,  
  main ="Title",  
  automatic_main = FALSE,  
  xlab = "x-axis",  
  ylab ="y-axis",  
  ylim = c(0,1),
  single=FALSE
  )

}
