% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_density_smooth.R
\name{smooth.construct.md.smooth.spec}
\alias{smooth.construct.md.smooth.spec}
\title{P-Splines with integrate-to-zero constraint for density regression}
\usage{
\method{smooth.construct}{md.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth specification object, usually generated by a term \code{ti(x,bs="md",...)}}

\item{data}{a list containing just the data (including any \code{by} variable) required by this term, 
            with names corresponding to \code{object$term} (and \code{object$by}). The \code{by} variable 
            is the last element.}

\item{knots}{a list containing any knots supplied for basis setup --- in same order and with same names as \code{data}. 
             Can be \code{NULL}. See details for further information.}
}
\value{
An object of class \code{"mdpline.smooth"}.
% See \code{\link[mgcv]{smooth.construct}}, for the elements that this object will contain.
}
\description{
Specifying \code{bs = "md"} in the tensor product smooth \code{\link[mgcv]{ti}}
(setting \code{mc = FALSE}) constructs a smoother to be used in \code{mgcv}'s
\code{\link[mgcv]{gam}} with \code{family = poisson()} for the case that the
response variable of interest is a density in a mixed Bayes Hilbert space
\eqn{B^2(\mu) = B^2(\mathcal{T}, \mathcal{A}, \mu)}{B^2(\mu) = B^2(T, A, \mu)}
(including continuous or discrete ones as special cases) as in Maier (2023).
The subset of the domain \eqn{\mathcal{T}}{T} corresponding to the continuous
part of the densities is denoted with \eqn{I}, the one corresponding to the
discrete part with \eqn{\mathcal{D}}{D}.
Corresponding covariate observations are the discrete values for the discrete
component and the midpoints of the bins underlying the histograms approximating
the densities for the continuous component. Response observations are the
counts of the histograms.
Specification of basis and penalization details for the continuous component
is as for \code{bs = "ps"}, see \code{mgcv}'s
\code{\link[mgcv]{smooth.construct.ps.smooth.spec}}.
Specification of the domain and for the discrete component is possible via
the argument \code{xt} in \code{\link[mgcv]{ti}} (see below). % \cr % Linebreak

Note that the implementation includes the integrate-to-zero constraint of
\eqn{L^2_0}, i.e., further centering is not reasonable! This means, that the
constructor should not be used with \code{\link[mgcv]{s}} or
\code{\link[mgcv]{te}}, which always include centering (sum-to-zero) constraints
for all marginals. Instead, it has to be used with \code{\link[mgcv]{ti}}, setting
\code{mc = FALSE} for the corresponding component!
% (It would be desirable to include a warning, however, object does not seem
% to contain information on which kind of function it was called in.)
}
\details{
The basis and penalty are constructed from a P-spline basis transformed to
\eqn{L^2_0} as described in Appendix D of Maier et al. (2021).
The argument \code{xt} in \code{\link[mgcv]{ti}} is used for further specification
regarding the underlying Bayes Hilbert space. \code{xt} has to be a list of
the same length as the vector \code{bs} specifying the marginal bases. For all
\code{"md"} type marginal bases, the corresponding \code{xt}-list element is
again a list with the following elements:
% ti(..., xt = list(list(values_discrete = NULL,
%                        weights_discrete = NULL,
%                        domain_continuous = NULL,
%                        penalty_discrete = NULL)))
\itemize{
\item
\code{values_discrete}: Vector of values in \eqn{\mathcal{D}}{D} (the subset of
the domain corresponding to the discrete part of the densities). Defaults to
missing (\code{NULL}) in which case it is set to \code{c(0, 1)}. If set to
\code{FALSE}, the discrete component is considered to be empty, i.e., the
Lebesgue measure is used as reference measure (continuous special case).
\item
\code{weights_discrete}: Vector of weights for the Dirac measures corresponding
to \code{values_discrete}. If missing (\code{NULL}) it is set to 1 in all
components as default. Can be a scalar for equal weights for all discrete values
or a vector with specific weights for each corresponding discrete value.
\item
\code{domain_continuous}: An interval (i.e., a vector of length 2) specifying
\eqn{I}. If missing (\code{NULL}) it is set to \code{c(0, 1)} as default. If
set to \code{FALSE}, the continuous component is considered to be empty, i.e.,
a weighted sum of dirac measures is used as reference measure (discrete special
case).
\item
\code{penalty_discrete}: integer (or \code{NULL}) giving the order of
differences to be used for the penalty of the discrete component, with \code{0}
corresponding to the identity matrix as penalty matrix (analogously to \code{m[2]}
for the continuous component); Note that the order of differences must be smaller
than the number of values in the discrete component, i.e., the length of
\code{values_discrete}; if missing (\code{NULL}), in the mixed case, it is
set to a zero matrix (corresponding to no penalization), while in the discrete
case, it is set to a diagonal matrix (as in a ridge penalty). In the discrete
case, please set the corresponding argument of \code{sp} in \code{\link[mgcv]{ti}}
to \code{0} to estimate an unpenalized model.
% For marginal bases of type \code{"ps"}/\code{"bs"}, the corresponding xt-list
% element has to be set to \code{NULL}.
}
}
\references{
Maier, E. M., Fottner, A., Stoecker, A., Okhrin, Y., & Greven, S. (2023):
Conditional density regression for individual-level data.

Maier, E.-M., Stoecker, A., Fitzenberger, B., Greven, S. (2021):
Additive Density-on-Scalar Regression in Bayes Hilbert Spaces with an Application to Gender Economics.
arXiv preprint arXiv:2110.11771.
}
\author{
Eva-Maria Maier, Alexander Fottner
}
