% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens_reg.R
\encoding{UTF-8}
\name{plot.dens_reg_obj}
\alias{plot.dens_reg_obj}
\title{Plot function for conditional density regression models}
\usage{
\method{plot}{dens_reg_obj}(
  obj,
  type = "histo",
  interactive = FALSE,
  level = "pdf",
  display_all = TRUE,
  pick_sites = FALSE,
  predict = NULL,
  terms = NULL,
  ...
)
}
\arguments{
\item{obj}{\code{dens_reg_obj}-object, i.e. the output of the \code{dens_reg}-function.}

\item{type}{"histo" or "effects": If \code{type="histo"}, the underlying histogram and the estimated conditional density is plotted for each unique covariate combination. If \code{type="effects"} the different partial effects are plotted.}

\item{interactive}{Indicates for \code{type="effects"} wether the plots are displayed as an interactive plot (\code{TRUE}) or if all plots are plotted consecutively.}

\item{level}{Indicates for \code{type="effects"} with "clr" that the clr-transformed effects in \eqn{L^2_0(\mu)} and with "pdf" that the partial effects in \eqn{B^2(\mu)} are plotted.}

\item{display_all}{Indicates for \code{type="effects"} and \code{predict=NULL} wether all curves are plotted for effects depending on continuous covariates, i.e., for each observed value of the covariate one curve is plotted, (\code{display_all=TRUE}) or if only three curves (min, med, max) are shown (\code{display_all=FALSE}).
If the effects are predicted for new data, i.e., \code{predict} is not \code{NULL}, \code{display_all=FALSE} causes that only the predicted effects for the first, middle and last row are displayed.}

\item{pick_sites}{Vector of integers from 1 to 6. Indicates for \code{type="effects"} if only certain effects should be plotted. Number corresponds to the order of the different effect types in the argument of \code{dens_reg}. If missing (\code{NULL}) all included effects are plotted.}

\item{predict}{If plots for certain covariate values are desired, \code{predict} has to be a data frame with new data in form of a data table with columns named as the relevant covariates. In each row, the user can specify a value of the respective covariate.}

\item{terms}{If predict is not \code{NULL}. Vector of term names or the indices of the terms (starting with 1=intercept) specifying which terms should be predicted and plotted.}
}
\value{
Plot(s) as specified.
}
\description{
\code{plot.dens_reg_obj} is the default plot method for data of the class \code{dens_reg_obj}.
}
\examples{

\donttest{# please run the examples of dens_reg to estimate the needed models

example("dens_reg")

#' # create new data for predict

nd<-data.frame(covariate1=c("a","b","c","a"),covariate4=c(0.4,0.5,0.1,0.3), covariate2=c("d","d","c","d"),covariate3=c(1,0,0.2,2),covariate5=c(0.2,0.4,1,2))


# plot mixed model (default settings: histogram, not interactive, density-level, display all)

plot(m_mixed)

# plot partial effects on clr-level of the continuous model in an interactive plot, do not show all groups

plot(m_cont, type="effects", interactive=TRUE, level="clr", display_all=FALSE)

# show only first plot

plot(m_cont, type="effects", interactive=FALSE, pick_sites=1, level="clr", display_all=FALSE)

# plot partial effects on density-level estimated for new data based on the mixed model

plot(m_mixed, type="effects", level="pdf", display_all=TRUE,predict=nd)

# estimate and plot only the intercept (first term)

plot(m_mixed, type="effects", level="pdf", display_all=TRUE,predict=nd, terms=1)

#' # estimate and plot only second term

plot(m_mixed, type="effects", level="pdf", display_all=TRUE,predict=nd, terms=2)

# customize some plot parameters, other plot parameters can also be specified by the user

plot(m_mixed, type="effects", display_all = FALSE,level="clr", main="Your title",legend.position="your legend", xlab="your xlab")
}

}
