% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens_reg.R
\name{predict.dens_reg_obj}
\alias{predict.dens_reg_obj}
\title{Predict function for conditonal density regression models}
\usage{
\method{predict}{dens_reg_obj}(
  obj,
  new_data = NULL,
  which = NULL,
  exclude = NULL,
  type = "terms",
  level = "clr"
)
}
\arguments{
\item{obj}{\code{dens_reg_obj}-object, i.e. the output of the \code{dens_reg}-function.}

\item{new_data}{New data in form of a data table with columns named as the relevant covariates for the terms which should be predicted. In each row, the user can specify a value of the respective covariate. If not specyfied (\code{NULL}) the values of the original data set (see obj$histo_data) are used for the predictions.}

\item{which}{Only terms (or their index number) named in this array will be predicted. Covariates only needed for these terms have to be given in \code{new_data}. If \code{which} and \code{exclude} are given, \code{exclude} will be used.}

\item{exclude}{Terms (or their index number) named in this array will not be predicted. Covariates only needed for terms which are excluded do not have to be given in \code{new_data}. If \code{which} and \code{exclude} are given, \code{exclude} will be used.}

\item{type}{"terms", "pdf" or "clr": If \code{type="terms"} each component of the linear predictor is returned seperately, if}

\item{level}{If \code{type="terms"}: "pdf" results in predicted terms on pdf-level, "clr" in terms on clr-level.  \code{type="pdf"}  or \code{type="clr"} specifies the prediction of \eqn{\hat f} or \eqn{clr(\hat f)} of the respective covariate values.}
}
\value{
A list of matrices (if type="terms") with one matrix for each term, different columns for every predicted covariate value.
A matrix with columns for the different covariate combinations if \code{type="pdf"} or \code{="clr"} containing the estimated \eqn{\hat f} or \eqn{clr(\hat f)}.
}
\description{
\code{predict.dens_reg_obj} is the default predict method for data of the class \code{dens_reg_obj}.
}
\examples{
\donttest{# please run the examples of ?dens_reg to estimate the needed models

# see names of the models' terms
sapply(m_mixed$model$smooth, "[[",  "label")
sapply(m_dis$model$smooth, "[[",  "label")
sapply(m_cont$model$smooth, "[[",  "label")

# create newdata for predict

nd<-data.frame(covariate1=c("a","b","c","a"),covariate4=c(0.4,0.5,0.1,0.3), covariate2=c("d","d","c","d"),covariate3=c(1,0,0.2,2),covariate5=c(0.2,0.4,1,2))

# predict mixed model, all terms on pdf-level without new data
p1<-predict(m_mixed, type= "terms",level="pdf")

# predict  partial effects at clr-level of the mixed model for new data
p2<-predict(m_mixed, type= "terms",  new_data=nd)

# predict only the second partial effect at density-level of the mixed model for new data
p3<-predict(m_mixed, type= "terms", which=all_terms[2], new_data=nd, level="pdf")

# predict partial effects of the mixed model for new data and at density-level without the second term
p4<-predict(m_mixed, type= "terms", exclude=all_terms[2], new_data=nd, level="pdf")

# predict f_hat on density-level for new data
p5<-predict(m_mixed, type= "pdf", new_data=nd)

# predict clr(f_hat) for new data
p6<-predict(m_mixed, type= "clr", new_data=nd)

}

}
